Restructure code into more appropriate modules
