+++ src/main/kotlin/com/github/derg/transpiler/phases/lexer/Tokenizer.kt
@@ -1,7 +1,8 @@
-package com.github.derg.transpiler.lexer
+package com.github.derg.transpiler.phases.lexer
 
 import com.github.derg.transpiler.core.Localized
 import com.github.derg.transpiler.core.Location
+import com.github.derg.transpiler.source.lexeme.*
 import com.github.derg.transpiler.util.indexOfFirstOrNull
 import com.github.derg.transpiler.util.indexOfOrNull
 import com.github.derg.transpiler.util.substringFrom
