+++ src/main/kotlin/com/github/derg/transpiler/phases/parser/ParserStatements.kt
@@ -1,12 +1,9 @@
-package com.github.derg.transpiler.parser.patterns
+package com.github.derg.transpiler.phases.parser
 
-import com.github.derg.transpiler.ast.*
 import com.github.derg.transpiler.core.Name
-import com.github.derg.transpiler.lexer.SymbolType
-import com.github.derg.transpiler.lexer.Token
-import com.github.derg.transpiler.parser.ParseError
-import com.github.derg.transpiler.parser.ParseOk
-import com.github.derg.transpiler.parser.Parser
+import com.github.derg.transpiler.source.ast.*
+import com.github.derg.transpiler.source.lexeme.SymbolType
+import com.github.derg.transpiler.source.lexeme.Token
 import com.github.derg.transpiler.util.Result
 
 /**
