+++ src/main/kotlin/com/github/derg/transpiler/phases/parser/ParserTokens.kt
@@ -1,13 +1,10 @@
-package com.github.derg.transpiler.parser.patterns
+package com.github.derg.transpiler.phases.parser
 
 import com.github.derg.transpiler.core.Name
-import com.github.derg.transpiler.lexer.Identifier
-import com.github.derg.transpiler.lexer.Symbol
-import com.github.derg.transpiler.lexer.SymbolType
-import com.github.derg.transpiler.lexer.Token
-import com.github.derg.transpiler.parser.ParseError
-import com.github.derg.transpiler.parser.ParseOk
-import com.github.derg.transpiler.parser.Parser
+import com.github.derg.transpiler.source.lexeme.Identifier
+import com.github.derg.transpiler.source.lexeme.Symbol
+import com.github.derg.transpiler.source.lexeme.SymbolType
+import com.github.derg.transpiler.source.lexeme.Token
 import com.github.derg.transpiler.util.Result
 import com.github.derg.transpiler.util.failureOf
 import com.github.derg.transpiler.util.successOf
