+++ src/test/kotlin/com/github/derg/transpiler/phases/parser/TestParserDefinitions.kt
@@ -1,9 +1,9 @@
-package com.github.derg.transpiler.parser.patterns
+package com.github.derg.transpiler.phases.parser
 
-import com.github.derg.transpiler.ast.Mutability
-import com.github.derg.transpiler.ast.Visibility
-import com.github.derg.transpiler.lexer.EndOfFile
-import com.github.derg.transpiler.parser.*
+import com.github.derg.transpiler.phases.parser.*
+import com.github.derg.transpiler.source.ast.Mutability
+import com.github.derg.transpiler.source.ast.Visibility
+import com.github.derg.transpiler.source.lexeme.EndOfFile
 import org.junit.jupiter.api.Test
 
 /**
